Consolidated Stimmkreis API calls to own file (#26)
