+++ frontend/src/components/stimmkreise/StimmkreisSelector.tsx
@@ -1,11 +1,6 @@
 import React from "react";
 import {ListGroup} from "react-bootstrap";
-
-export interface StimmkreisInfo {
-    id: number;
-    name: string;
-    number: number;
-}
+import StimmkreisEndpoints, {StimmkreisInfo} from "../../rest_client/StimmkreisEndpoints";
 
 interface Props {
     filterText?: string,
@@ -28,18 +23,15 @@ export class StimmkreisSelector extends React.Component<Props> {
     }
 
     componentDidMount() {
-        console.log('Fetching');
-        fetch('/api/stimmkreise')
-            .then(response => response.json())
-            .then(data => {
-                this.setState({
-                    stimmkreisInfo: data,
-                });
-                // Select the first Stimmkreis by default
-                if (data.length) {
-                    this.setSelection(data[0]);
-                }
-            })
+        StimmkreisEndpoints.getAllInfo().then(data => {
+            this.setState({
+                stimmkreisInfo: data,
+            });
+            // Select the first Stimmkreis by default
+            if (data.length) {
+                this.setSelection(data[0]);
+            }
+        })
     }
 
     performFilter(stimmkreise: StimmkreisInfo[], filterTerm: string|undefined) : StimmkreisInfo[] {
